finalização

+++ Entrega 19-04/alg158.py
@@ -1,4 +1,4 @@
-x = int(input('Entre com o valor de x: '))
+x = float(input('Entre com o valor de x: '))
 if x != 2:
     fx = 8/(2-x)
     print('Fx = {}'.format(fx))
