+++ Entrega 19-04/alg159.py
@@ -1,4 +1,4 @@
-x = int(input('Entre com o valor de x: '))
+x = float(input('Entre com o valor de x: '))
 if x > 4 or x < (-4):
     fx = (5*x+3)/((x**2 - 16)**0.5)
     print('Fx = {}'.format(fx))
